Crate flux_metadata

Modules

decoder 🔒

encoder 🔒

errors 🔒

Macros

get 🔒

Structs

AbsoluteBytePos

CStore

CrateMetadata

EncodedSourceFileId 🔒

From CREUSOT: used to store the info about source files An EncodedSourceFileId is the same as a StableSourceFileId except that the source crate is represented as a StableCrateId instead of as a CrateNum. This way EncodedSourceFileId can be encoded and decoded without any additional context, i.e. with a simple opaque::Decoder (which is the only thing available when decoding the Footer.

Footer

SourceFileIndex 🔒

Tables

Constants

METADATA_HEADER 🔒

METADATA_VERSION 🔒

SYMBOL_OFFSET 🔒

SYMBOL_PREDEFINED 🔒

SYMBOL_STR 🔒

TAG_FULL_SPAN 🔒

TAG_PARTIAL_SPAN 🔒

Statics

DEFAULT_LOCALE_RESOURCE

Raw content of Fluent resource for this crate, generated by fluent_messages macro, imported by rustc_driver to include all crates’ resources in one bundle.

Traits

Key 🔒

Trait to deal with the fact that assoc_refinmenents_of and assoc_refinements_def use FluxId<K> as key;

Functions

encode_def_ids 🔒

encode_flux_defs 🔒

encode_metadata

filename_for_metadata

flux_metadata_extern_location 🔒